Fix setup schema alignment and gate coverage

+++ crates/db/src/repo/idempotency.rs
@@ -7,7 +7,7 @@ pub async fn store(
     endpoint: &str,
     payload_hash: &str,
     response: &str,
-    status_code: i64,
+    status_code: i32,
 ) -> Result<(), sqlx::Error> {
     let now = chrono::Utc::now().timestamp();
     sqlx::query(
@@ -31,13 +31,13 @@ pub struct IdempotencyRecord {
     pub endpoint: String,
     pub payload_hash: String,
     pub response: String,
-    pub status_code: i64,
+    pub status_code: i32,
     pub created_at: i64,
 }
 
 /// Lookup an existing idempotency key.
 pub async fn lookup(pool: &DbPool, key: &str) -> Result<Option<IdempotencyRecord>, sqlx::Error> {
-    let row: Option<(String, String, String, String, i64, i64)> = sqlx::query_as(
+    let row: Option<(String, String, String, String, i32, i64)> = sqlx::query_as(
         "SELECT key, endpoint, payload_hash, response, status_code, created_at FROM idempotency_keys WHERE key = $1",
     )
     .bind(key)
